moved parent / window to folder and renamed window to global.
