+++ src/parent/index.ts
@@ -1,26 +1,11 @@
-import { BankPacksInfos } from "./bank";
-import { CharacterEntity } from "./entities/character-entity";
-import { MonsterEntity } from "./entities/monster-entity";
-import { NpcEntity } from "./entities/npc-entity";
-import { Entity } from "./entity";
-import { PartyCharacter } from "./functions";
-import {
-  MapKey,
-  GDropList,
-  GDropMaps,
-  GDropMonsters,
-  GDropNormalDrops,
-  ItemKey,
-  MonsterKey,
-  EventKey,
-} from "./G";
-import { PositionReal } from "./position";
-import { ServerIdentifier, ServerRegion } from "./server";
-import { SocketWithEventsFunctions } from "./socket-events";
-import { SkillKey } from "./types/GTypes/skills";
-import { BetterUXWrapper } from "./types/GTypes/utils";
-
-export * from "./socket-events";
+import { BankPacksInfos } from "../bank";
+import { CharacterEntity, MonsterEntity, NpcEntity } from "../entity";
+import { PartyCharacter } from "../functions";
+import { EventKey, GDropList, GDropMaps, GDropMonsters, GDropNormalDrops, ItemKey, MapKey, MonsterKey, SkillKey } from "../G";
+import { PositionReal } from "../position";
+import { ServerIdentifier, ServerRegion } from "../server";
+import { SocketWithEventsFunctions } from "../socket-events";
+import { BetterUXWrapper } from "../types/GTypes/utils";
 
 /** Tracktrix informations */
 export interface Tracker {
@@ -175,8 +160,9 @@ export type SEventsInfos = {
   SPartialABTesting &
   SPartialGooBrawlEventInfo;
 
-export {}; // this is done to make window a module
+export {}; // this is done to make parent a module
 declare global {
+  /** When you access parent via game code, this is what you have access to. */
   interface Window {
     $: JQueryStatic;
     clear_game_logs(): void;
@@ -237,47 +223,6 @@ declare global {
 
     S: SEventsInfos;
   }
-
-  /* eslint-disable no-var, vars-on-top */
-  var handle_command: undefined | ((command: string, args: string) => void);
-  var on_cm: undefined | ((from: string, data: any) => void);
-  // var on_map_click: undefined | ((x: number, y: number) => boolean);
-
-  /**
-   *
-   */
-  var on_party_invite: undefined | ((from: string) => void);
-
-  /**
-   * called by the inviter's name
-   * request = someone requesting to join your existing party
-   */
-  var on_party_request: undefined | ((from: string) => void);
-
-  var on_disappear: undefined | ((entity: Entity, data: unknown) => void);
-
-  /**
-   * called by the mage's name in PVE servers,
-   * in PVP servers magiport either succeeds or fails without consent
-   */
-  var on_magiport: undefined | ((from: string) => void);
-
-  /**
-   * if true is returned, the default move is cancelled
-   */
-  var on_map_click: undefined | ((x: number, y: number) => void);
-
-  /**
-   * called just before the CODE is destroyed
-   */
-  var on_destroy: undefined | (() => void);
-
-  /**
-   * the game calls this function at the best place in each game draw frame,
-   * so if you are playing the game at 60fps, this function gets called 60 times per second
-   */
-  var on_draw: undefined | (() => void);
-  /* eslint-enable no-var, vars-on-top */
 }
 
 export type ChestInfo = PositionReal & {
