+++ src/socket-events/server-to-client/events/ServerToClient_tracker.ts
@@ -1,3 +1,3 @@
-import { Tracker } from "../../../window";
+import { Tracker } from "../../../parent";
 
 export type ServerToClient_tracker = Tracker;
